backend/pyproject.toml
[project]
name = "books"
version = "1.0.0"
description = "The basic blog app built in the Flask tutorial."
requires-python = ">=3.7"
dependencies = [
    "flask",
    "Flask-Alembic",
    "Flask-SQLAlchemy",
    "psycopg2-binary"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
